requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.0.1
lxml==4.6.3
Markdown==3.3.4
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
